+++ lib/angelo/params_parser.rb
@@ -62,8 +62,19 @@ module Angelo
       super(){|hash,key| hash[key.to_s] if Symbol === key}
       unless h.nil?
         merge! h
+
         # Replace values that are Hashes with SymHashes, recursively.
-        each {|k,v| self[k] = SymHash.new(v) if Hash === v}
+        each do |k,v|
+          self[k] = case v
+                    when Hash
+                      SymHash.new(v)
+                    when Array
+                      v.map {|e| Hash === e ? SymHash.new(e) : e}
+                    else
+                      v
+                    end
+        end
+
       end
     end